requirements.txt
numpy==1.24.3
pandas==2.0.3
transformers==4.31.0
datasets==2.13.1
torch==2.0.1
scikit-learn==1.3.0
tqdm==4.65.0
pyyaml==6.0.1
jsonlines==3.1.0
nltk==3.8.1
spacy==3.6.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
sentencepiece==0.1.99
huggingface-hub==0.16.4
accelerate==0.21.0 